Copter: interpret input in stabilize as a tilt vector

+++ ArduCopter/control_autotune.pde
@@ -249,7 +249,7 @@ static bool autotune_start(bool ignore_checks)
 // should be called at 100hz or more
 static void autotune_run()
 {
-    int16_t target_roll, target_pitch;
+    float target_roll, target_pitch;
     float target_yaw_rate;
     int16_t target_climb_rate;
